+++ examples/shardingsphere-example-generator/src/main/java/org/apache/shardingsphere/example/generator/core/impl/JDBCExampleGenerator.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.example.generator.core.impl;
 import freemarker.template.Configuration;
 import freemarker.template.TemplateException;
 import org.apache.shardingsphere.example.generator.core.ExampleGenerator;
-import org.apache.shardingsphere.example.generator.core.GenerateUtil;
+import org.apache.shardingsphere.example.generator.core.GenerateUtils;
 import org.apache.shardingsphere.example.generator.scenario.ExampleScenarioFactory;
 
 import java.io.IOException;
@@ -38,12 +38,12 @@ public final class JDBCExampleGenerator implements ExampleGenerator {
     @Override
     public void generate(final Configuration templateConfig, final Map<String, String> dataModel, final String relativePath) throws IOException, TemplateException {
         ExampleScenarioFactory exampleScenarioFactory = new ExampleScenarioFactory(dataModel.get("feature"), dataModel.get("framework"), dataModel.get("transaction"));
-        GenerateUtil.generateDirs(templateConfig, dataModel, exampleScenarioFactory.getJavaClassPaths(), relativePath + JAVA_CLASS_PATH);
-        GenerateUtil.generateDirs(templateConfig, dataModel, exampleScenarioFactory.getResourcePaths(), relativePath + RESOURCES_PATH);
-        GenerateUtil.generateFile(templateConfig, getType(), dataModel, exampleScenarioFactory.getJavaClassTemplateMap(), relativePath + JAVA_CLASS_PATH);
-        GenerateUtil.generateFile(templateConfig, getType(), dataModel, exampleScenarioFactory.getResourceTemplateMap(), relativePath + RESOURCES_PATH);
-        String outputPath = GenerateUtil.generatePath(templateConfig, dataModel, relativePath);
-        GenerateUtil.processFile(templateConfig, dataModel, getType() + "/pom.ftl", outputPath + "pom.xml");
+        GenerateUtils.generateDirs(templateConfig, dataModel, exampleScenarioFactory.getJavaClassPaths(), relativePath + JAVA_CLASS_PATH);
+        GenerateUtils.generateDirs(templateConfig, dataModel, exampleScenarioFactory.getResourcePaths(), relativePath + RESOURCES_PATH);
+        GenerateUtils.generateFile(templateConfig, getType(), dataModel, exampleScenarioFactory.getJavaClassTemplateMap(), relativePath + JAVA_CLASS_PATH);
+        GenerateUtils.generateFile(templateConfig, getType(), dataModel, exampleScenarioFactory.getResourceTemplateMap(), relativePath + RESOURCES_PATH);
+        String outputPath = GenerateUtils.generatePath(templateConfig, dataModel, relativePath);
+        GenerateUtils.processFile(templateConfig, dataModel, getType() + "/pom.ftl", outputPath + "pom.xml");
     }
     
     @Override
